refactor: make shake for support a native bottom sheet
